3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -1,5 +1,5 @@
 #include "variadic_functions.h"
-#include <starg.h>
+#include <stdarg.h>
 #include <stdio.h>
 
 /**
@@ -28,7 +28,7 @@ void print_all(const char * const format, ...)
 					printf("%s%d", sep, va_arg(list, int));
 					break;
 				case 'f':
-					printf("%s%f", sep, va_arg(list, int));
+					printf("%s%f", sep, va_arg(list, double));
 					break;
 				case 's':
 					str = va_arg(list, char*);
